refactor: extract VerificationCardError enum to standalone module
Move the enum to a lightweight file without heavy dependencies so test
files that mock @bifold/core and @react-navigation/native can import it
safely. Use enum instead of string literal in useCardScanner test.
