+++ app/src/bcsc-theme/features/verify/verificationCardError.ts
@@ -1,0 +1,4 @@
+export enum VerificationCardError {
+  MismatchedSerial = 'MismatchedSerial',
+  CardExpired = 'CardExpired',
+}
