+++ app/src/bcsc-theme/hooks/useCardScanner.test.tsx
@@ -1,6 +1,7 @@
 import useApi from '@/bcsc-theme/api/hooks/useApi'
 import { useCardScanner } from '@/bcsc-theme/hooks/useCardScanner'
 import { useSecureActions } from '@/bcsc-theme/hooks/useSecureActions'
+import { VerificationCardError } from '@/bcsc-theme/features/verify/verificationCardError'
 import { BCSCScreens } from '@/bcsc-theme/types/navigators'
 import { ScanableCode } from '@/bcsc-theme/utils/decoder-strategy/DecoderStrategy'
 import * as Bifold from '@bifold/core'
@@ -385,7 +386,7 @@ describe('useCardScanner', () => {
           { name: BCSCScreens.SetupSteps },
           {
             name: BCSCScreens.VerificationCardError,
-            params: { errorType: 'MismatchedSerial' },
+            params: { errorType: VerificationCardError.MismatchedSerial },
           },
         ],
       })
