+++ app/src/bcsc-theme/types/navigators.ts
@@ -1,6 +1,6 @@
 import { NavigatorScreenParams } from '@react-navigation/native'
 import { BCSCCardProcess, EvidenceType } from 'react-native-bcsc-core'
-import { VerificationCardError } from '../features/verify/VerificationCardErrorScreen'
+import { VerificationCardError } from '../features/verify/verificationCardError'
 import { BCSCReason } from '../utils/id-token'
 
 export enum BCSCStacks {
